working red mid, pid retune, archival
