+++ MeepMeepTesting/src/main/java/com/example/meepmeeptesting/RedRight50rr.java
@@ -17,8 +17,8 @@ public class RedRight50rr { //closest to backdrop
                         drive.trajectorySequenceBuilder(new Pose2d(12, -63, Math.toRadians(0)))
                                 // .splineTo(new Vector2d(12, 34), Math.toRadians(90))
                                 // .splineTo(new Vector2d(12, 52), Math.toRadians(90))
-                                .splineTo(new Vector2d(22, -37), Math.toRadians(90))
-                                .back(5)
+                                .splineTo(new Vector2d(22, -42), Math.toRadians(90))
+                                .splineToConstantHeading(new Vector2d(22, -52), Math.toRadians(90))
                                 .splineToSplineHeading(new Pose2d(44.5, -36, Math.toRadians(0)), Math.toRadians(0))
                                 //.splineTo(new Vector2d(44.5, -36), Math.toRadians(180))
                                 .strafeRight(20) // separate trajectory
