+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/AutoUtil.java
@@ -1,13 +1,21 @@
 package org.firstinspires.ftc.teamcode.util;
 
+import static android.os.SystemClock.sleep;
+
+import com.arcrobotics.ftclib.controller.PIDController;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
+import com.qualcomm.robotcore.hardware.Servo;
 
 import static org.firstinspires.ftc.teamcode.util.RobotConstants.*;
 
 public class AutoUtil {
     public LinearOpMode opMode;
     public DcMotor frontLeft, frontRight, backLeft, backRight, arm, joint;
+    public Servo leftClaw, rightClaw;
+
+    private PIDController armPID = new PIDController(armP, armI, armD);
+    private PIDController jointPID = new PIDController(jointP, jointI, jointD);
 
     public enum RobotState {
         IDLE,
@@ -23,8 +31,9 @@ public class AutoUtil {
 
     public RobotState currentState = RobotState.IDLE;
 
+
     // Constructor
-    public AutoUtil(LinearOpMode opMode, DcMotor frontLeft, DcMotor frontRight, DcMotor backLeft, DcMotor backRight, DcMotor arm, DcMotor joint) {
+    public AutoUtil(LinearOpMode opMode, DcMotor frontLeft, DcMotor frontRight, DcMotor backLeft, DcMotor backRight, DcMotor arm, DcMotor joint, Servo leftClaw, Servo rightClaw) {
         this.opMode = opMode;
         this.frontLeft = frontLeft;
         this.frontRight = frontRight;
@@ -32,6 +41,8 @@ public class AutoUtil {
         this.backRight = backRight;
         this.arm = arm;
         this.joint = joint;
+        this.leftClaw = leftClaw;
+        this.rightClaw = rightClaw;
     }
 
     public void waitForDrive() {
@@ -164,6 +175,39 @@ public class AutoUtil {
         opMode.telemetry.addData(name + " Power", motor.getPower());
         opMode.telemetry.addData(name + " Position", motor.getCurrentPosition());
         opMode.telemetry.addData(name + " Target Position", motor.getTargetPosition());
+    }
+
+    public void moveArm(double target) {
+        PIDController armPID = new PIDController(armP, armI, armD);
+        double error = target - arm.getCurrentPosition();
+
+        while (opMode.opModeIsActive() && Math.abs(error) > 10) {
+            // calculate angles of joint & arm (in degrees) to account for torque
+            double joint_angle = 193;
+            double relative_arm_angle = arm.getCurrentPosition() / RobotConstants.arm_ticks_per_degree + 14.8;
+            double arm_angle = 270 - relative_arm_angle - joint_angle;
+
+            double arm_ff = Math.cos(Math.toRadians(arm_angle)) * armF;
+
+            error = target - arm.getCurrentPosition();
+
+            double arm_out = armPID.calculate(arm.getCurrentPosition(), target);
+
+            double arm_power = arm_ff + arm_out;
+
+            arm.setPower(arm_power);
+
+            motorTelemetry(arm, "Arm");
+            opMode.telemetry.addData("Error", error);
+            opMode.telemetry.addData("Power", arm_power);
+            opMode.telemetry.update();
+            sleep(5);
+        }
+
+    }
+
+    public void moveLeftFinger(double target) {
+        leftClaw.setPosition(target);
     }
 
 }
